fix: get build info only in ckb executable crate.
The build info relies on crate build script. If we put the build script in
a sub-crate, the build info is only refreshed when that crate and its
dependencies have changed, which may be out-of-date. Only the build script in
the final executable crate can ensure to be the latest.