feat: Include fractions in epoch number representations
This change introduce fractions in 2 places where epoch numbers might
be used:

* The epoch field in the header
* Cell input's since part when using epoch values

Specifically, the new epoch number format works as follows:

* Only the lower 56 bits of the full 64-bit value are used
* The upper 40 bits are used to represent the current epoch number
* The lower 16 bits represent the already elapsed fractions of current
  epoch(not including current block)

Assuming we are at block number 11555, epoch 50, and epoch 50 starts
from block 11000, has a length of 1000. The epoch number for this
particular block will then be 3313172, which is calculated in the
following way:

```
(50 << 16) | (((2 ** 16) * (11555 - 11000) / 1000) & ((2 ** 16) - 1))
```

Note the formula uses normal integer division which rounded to the
floor.

This way, we can alleviate the trouble that CKB uses variable epoch
length. The actual contract or dapps can just read this epoch number
with fractions and perform numerical operations as normal.