feat: Switch to rational number based epoch number format
This commit changes the epoch number to a rational number based
format. The lower 54 bits of the epoch number field are split into 3
parts(listed in the order from higher bits to lower bits):

* The highest 16 bits represent the epoch length
* The next 16 bits represent the current block index in the epoch
* The lowest 24 bits represent the current epoch number

Different from previous solution, with a rational number format the
precision can be guaranteed.

Assuming we are at block number 11555, epoch 50, and epoch 50 starts
from block 11000, has a length of 1000. The epoch number for this
particular block will then be 9326559282, which is calculated
in the following way:

```
50 | ((11555 - 11000) << 24) | (1000 << 16)
```