feat: Reject new scripts with known bugs
For compatibility reasons, there're certain bugs that we have to leave
to the next hardfork to fix. However those bugs, especially VM bugs
might lead to surprising unexpected behaviors. This change adds a new
checker that checks against newly created cells for scripts with bugs,
and reject those transaction when we can. This way we can alert users
about the bugs as early as we can.