refactor: add a link on script exit code
Each script can define its own exit codes. Later, there may be a site
can search the exit code linked to the failed script. But now we only
have a table for the system scripts.

This commit adds the link to the system script exit codes table in the
error message..