doc: use community contributed site for script error codes
Before the error message only link to the error codes for the system
scripts. This commit links to a page organized by script hash type and
code hash. The community can contribute the error codes pages for
various scripts in https://github.com/nervosnetwork/ckb-script-error-codes

See the example page for the built-in `TYPE_ID` script:

https://nervosnetwork.github.io/ckb-script-error-codes/by-type-hash/00000000000000000000000000000000000000000000000000545950455f4944.html#-3