feat: Add flatmemory feature for FlatMemory based machine types
This change adds a new `flatmemory` feature to ckb-script, which will
use `FlatMemory` as the memory type for
`CoreMachine`/`CoreMachineType`. While this is not gonna be used in
CKB, a FlatMemory will be quite useful in the development of
surrounding tools, including ckb-debugger. Note that one option is
that a debugger could maintain its own ckb-script package, but
considering the fact that the change here is rather small, I would
suggest we include this here feature in CKB.