fix: Fix one regression caused by moving IO processing code
Earlier Meepo implementation has a bug: when IO processing code suspends
or resumes VMs, the cycles consumed by the suspending / resuming
operation will not be reflected in the `current cycles` syscalls made by
the subsequent VM execution. While the bug can be thought as a
misimplementation, it is causing regression when we are moving IO
processing code, and a script is using `current cycles` syscall for eal.