feat: bump crates MSRV to 1.92.0
MSRV is missing in crates Cargo.toml.

This commit also marks `feat:` so release-plz will bump the minor version for all crates. This should be done in #4993 but have not.

If you bump MSRV in a patch release (e.g., 1.1.1 → 1.1.2), users automatically get the broken build when they run cargo update, with no easy way to opt out.